requirements.txt
Appium-Python-Client==0.47
robotframework==3.1.2
robotlibcore-temp==1.0.2
